requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
Pillow==10.2.0